Add channel topics
A whole bunch of CSS shenanigans necessary to accommodate this --
eventually got the scrolling right with a bunch of absolute positioning.

Also got a box above the nicks for stats there (total in the channel etc.),
but it's hidden for now as there's no code yet to show the stats.